backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.36
requests==2.31.0
boto3==1.34.40
python-dotenv==1.0.1
